+++ src/graphql/schema.ts
@@ -7,7 +7,7 @@ export const typeDefs = `#graphql
     type Heartbeat {
         email_name: String!
         last_heartbeat: Int
-        disabled_until: Int
+        disabled_until: Float
         is_disabled: Boolean!
         always_forward: Boolean!
         max_heartbeat_interval_seconds: Int!
@@ -55,9 +55,9 @@ export const typeDefs = `#graphql
     }
     
     type Mutation {
-        createHeartbeat(email_name: String!, max_heartbeat_interval_seconds: Int!, matching_criteria: String!, endpoint_id: Int, forwarding_token: String, description: String, always_forward: Boolean, disabled_until: Int): Heartbeat
+        createHeartbeat(email_name: String!, max_heartbeat_interval_seconds: Int!, matching_criteria: String!, endpoint_id: Int, forwarding_token: String, description: String, always_forward: Boolean, disabled_until: Float): Heartbeat
         recordHeartbeat(email_name: String!): ConciseHeartbeat
-        updateHeartbeat(email_name: String!, max_heartbeat_interval_seconds: Int, matching_criteria: String, endpoint_id: Int, forwarding_token: String, description: String, always_forward: Boolean, disabled_until: Int): Heartbeat
+        updateHeartbeat(email_name: String!, max_heartbeat_interval_seconds: Int, matching_criteria: String, endpoint_id: Int, forwarding_token: String, description: String, always_forward: Boolean, disabled_until: Float): Heartbeat
         deleteHeartbeat(email_name: String!): Boolean
         createPushoverEndpoint(user_key: String!, timezone: String!, description: String): PushoverEndpoint
         updatePushoverEndpoint(id: Int!, user_key: String, timezone: String, description: String): PushoverEndpoint
